+++ packages/compiler/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @openfn/compiler
 
+## 0.0.12
+
+### Patch Changes
+
+- Updated dependencies
+  - @openfn/logger@0.0.5
+
 ## 0.0.11
 
 ### Patch Changes